backend/reward-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-dotenv==1.0.0
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.29.0
httpx==0.24.0
alembic==1.10.4
PyJWT==2.6.0
